Refactored files structure
